requirements.txt
Telethon
Jinja2==3.1.3
PyYAML==6.0.1
aiofiles==23.2.1
cryptg==0.4.0  # Optional but recommended for speed
hachoir==3.3.0
PySocks # For metadata extraction